fixed ErrorProtectionSpecificConfig parse

+++ libflv/source/mpeg4-hevc.c
@@ -264,7 +264,7 @@ int mpeg4_hevc_codecs(const struct mpeg4_hevc_t* hevc, char* codecs, size_t byte
     x = ((x >> 4) & 0x0f0f0f0f) | ((x & 0x0f0f0f0f) << 4);
     x = ((x >> 8) & 0x00ff00ff) | ((x & 0x00ff00ff) << 8);
     x = (x >> 16) | (x << 16);
-    return snprintf(codecs, bytes, "hvc1.%s%d.%x.%c%d", space[hevc->general_profile_space%4], hevc->general_profile_idc, x, tier[hevc->general_tier_flag%2], hevc->general_level_idc);
+    return snprintf(codecs, bytes, "hvc1.%s%u.%x.%c%u", space[hevc->general_profile_space%4], (unsigned int)hevc->general_profile_idc, (unsigned int)x, tier[hevc->general_tier_flag%2], (unsigned int)hevc->general_level_idc);
 }
 
 #if defined(_DEBUG) || defined(DEBUG)
